Replace UnicodeStringView with std::u32string_view.
This was a decision that was to be made eventually. UnicodeStringView
cannot support constant-time indexing by design, and indexing/slicing is
likely to be demanded by the library users. Now we force a UTF-32 memory
representation of the regex (thus likely an additional utf-to-utf
conversion), but this only seems to be a minor issue given that
copious memory allocations and unicode transcoding are already there.